pyproject.toml
[project]
name = "intel-3"
version = "0.1.0"
description = "docintel-ragbot"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "chromadb>=1.5.0",
    "docling>=2.73.0",
    "hf-xet>=1.2.0",
    "langchain-chroma>=1.1.0",
    "langchain-community>=0.4.1",
    "langchain-huggingface>=1.2.0",
    "langchain-openai>=1.1.9",
    "langchain-text-splitters>=1.1.0",
    "openai>=2.20.0",
    "pillow>=11.3.0",
    "python-dotenv>=1.2.1",
    "sentence-transformers>=5.2.2",
]
